fix: remove nvim-dap Session.connect patch
I monkeypatched Session.connect so that I could increase the retry
timeout when connecting to the DAP server since sometimes plz debug
would take too long to start up. This means having to maintain
functional parity between the monkeypatched Session.connect and the
actual implementation though which is proving to be more work than it's
worth as nvim-dap continues to add new features.